Fix FP32 mode numpy dtype

+++ anyvision_detector.py
@@ -265,7 +265,7 @@ def print_to_file(frame_num, f, bbs):
             w = X2 - X1
             h = Y2 - Y1
 
-            total_string += str(the_class) + "," + str(score) + "," + str(X1 + int(w / 2)) + "," + str(Y1 + int(h / 2)) + "," + str(w) + "," + str(h)
+            total_string += str(the_class) + "," + str(score) + "," + str(X1 + w / 2) + "," + str(Y1 + h / 2) + "," + str(w) + "," + str(h)
 
             if i != bb_num - 1:
                 total_string += " "
@@ -371,7 +371,7 @@ if __name__ == "__main__":
     else:
         args.two_patches = False
     ort_session = ort.InferenceSession("detector_fp16.onnx" if args.fp16 else "detector.onnx")
-    input_dtype = np.half if args.fp16 else np.float
+    input_dtype = np.half if args.fp16 else np.float32
     if isDirectory != True:
         print("working on vid:", args.data_path, "\n\n")
         create_detections(ort_session, args.data_path, input_dtype, args, two_patches=args.two_patches)
